+++ src/apiflask/route.py
@@ -76,9 +76,15 @@ def route_patch(cls):
         is_view_class: bool = False
 
         if hasattr(view_func, 'view_class'):
+            view_class = view_func.view_class  # type: ignore
             # a function returned by MethodViewClass.as_view()
             is_view_class = True
-            view_class = view_func.view_class  # type: ignore
+            if not isinstance(view_class, type(MethodView)):
+                # avoid recording spec for view class
+                is_view_class = False
+                # avoid auto response being added to bare view
+                view_func._spec = {'hide': True}  # type: ignore
+
         elif isinstance(view_func, type(MethodView)):
             # a MethodView class passed with the route decorator
             is_view_class = True