+++ CHANGES.md
@@ -11,12 +11,14 @@ Released: -
 - Change the status code of request validation error from 400 to 422 ([issue #345][issue_345]).
 - Improve the default bypassing rules to support bypass blueprint's static endpoint and
   Flask-DebugToolbar ([issue #344][issue_344], [issue #369][issue_369]).
+- Explicitly check if `view_func.view_class` is `MethodViewType` in `add_url_rule` ([issue #379][issue_379]).
 
 [issue_350]: https://github.com/apiflask/apiflask/issues/350
 [issue_349]: https://github.com/apiflask/apiflask/issues/349
 [issue_345]: https://github.com/apiflask/apiflask/issues/345
 [issue_344]: https://github.com/apiflask/apiflask/issues/344
 [issue_369]: https://github.com/apiflask/apiflask/issues/369
+[issue_379]: https://github.com/apiflask/apiflask/issues/379
 
 
 ## Version 1.1.3